@@ -8,9 +8,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using RazorPages.Pages;
 
-namespace RazorPages
-{
+namespace RazorPages;
+
     public class Startup
     {
         public Startup(IConfiguration configuration)
@@ -24,11 +25,15 @@ namespace RazorPages
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddRazorPages();
+            services.AddScoped<IPortfolioContentService, PortfolioContentService>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
+
+
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -53,4 +58,4 @@ namespace RazorPages
             });
         }
     }
-}
+
